feat: migrate from Nuxt 4 to Astro
Replace Nuxt 4 + @nuxt/content with Astro + content collections.

Nuxt Content ships a WASM SQLite database to the browser even for
fully static sites, and view transitions have been unreliable. Astro
content collections are build-time only (zero client JS) and
ClientRouter provides stable view transitions.

- Astro components for all static content (pages, layouts, cards)
- Vue islands for interactivity (DarkModeToggle, HomeAvatar, AppNowPlaying)
- Last.fm API endpoint as Cloudflare Worker via Astro API routes
- Tailwind CSS 4 + Radix Colors carried over unchanged
- astro-icon replaces @nuxt/icon (same Iconify ecosystem)
- Path-based pagination replaces query param pagination
- Remark plugin replaces Nuxt module for reading time
- Includes draft blog post documenting the migration
