+++ .gitignore
@@ -1,10 +1,12 @@
-# Nuxt dev/build outputs
+# Build outputs
+.astro
 .output
 .data
 .nuxt
 .nitro
 .cache
 dist
+.wrangler
 
 # Node dependencies
 node_modules
