--- app/components/app/AppFooter.vue
@@ -1,10 +1,0 @@
-<template>
-  <footer class="mt-16 w-full bg-sage-3 dark:bg-sagedark-2">
-    <div class="mx-auto grid max-w-6xl grid-cols-12 gap-4 px-4 py-12 md:gap-8 lg:gap-12">
-      <div class="col-span-12 flex flex-col gap-12 md:col-span-8 md:col-start-3">
-        <AppNowPlaying />
-        <AppSocialLinks />
-      </div>
-    </div>
-  </footer>
-</template>
