--- app/components/home/HomePage.vue
@@ -1,32 +1,0 @@
-<template>
-  <div class="col-span-12 md:col-span-8 md:col-start-3">
-    <div class="flex flex-col items-center gap-8 md:flex-row md:items-center">
-      <HomeAvatar class="aspect-square w-full max-w-48 shrink-0 md:max-w-52" />
-      <div class="grid auto-rows-auto content-start gap-4">
-        <div class="relative">
-          <h1
-            class="animate-[fadeInUp_0.8s_cubic-bezier(0.16,1,0.3,1)_forwards] font-serif text-[clamp(2.5rem,8vw,4rem)] leading-[1.1] font-light tracking-tighter text-balance text-jade-11 opacity-0 dark:text-jadedark-11"
-          >
-            I build interfaces that
-            <span class="italic">work</span>
-          </h1>
-        </div>
-
-        <div class="animate-[fadeInUp_0.8s_cubic-bezier(0.16,1,0.3,1)_0.5s_forwards] space-y-5 opacity-0">
-          <p class="text-xl leading-normal text-pretty text-sage-12 md:text-2xl dark:text-sagedark-11">
-            Hey, I'm Seán, a front-end engineer building
-            <span class="font-serif italic">consistent</span>
-            ,
-            <span class="font-serif italic">accessible</span>
-            and
-            <span class="font-serif italic">delightful</span>
-            interfaces for the people who use them.
-          </p>
-        </div>
-      </div>
-    </div>
-  </div>
-  <LazyHomeRecentPosts
-    class="col-span-12 animate-[fadeInUp_0.8s_cubic-bezier(0.16,1,0.3,1)_0.7s_forwards] opacity-0 md:col-span-8 md:col-start-3"
-  />
-</template>
