--- app/pages/index.vue
@@ -1,10 +1,0 @@
-<script setup lang="ts">
-useSeoMeta({
-  description:
-    "Seán O'Grady is a front-end engineer building consistent, accessible and delightful interfaces for the people who use them.",
-});
-</script>
-
-<template>
-  <HomePage />
-</template>
