--- app/plugins/luxon.ts
@@ -1,5 +1,0 @@
-import { Settings } from 'luxon';
-
-export default defineNuxtPlugin(() => {
-  Settings.defaultLocale = 'en-US';
-});
