+++ package.json
@@ -1,80 +1,43 @@
 {
-  "name": "nuxt-app",
+  "name": "seanogrady-me",
   "private": true,
   "type": "module",
   "scripts": {
-    "build": "nuxt build",
-    "check": "pnpm type-check && pnpm knip && pnpm lint",
-    "dev": "nuxt dev",
-    "format": "oxfmt",
-    "generate": "nuxt generate",
-    "knip": "knip",
-    "lint": "pnpm lint:oxlint && pnpm lint:eslint",
-    "lint:eslint": "eslint --fix --cache --concurrency auto",
-    "lint:oxlint": "oxlint --type-aware --fix",
-    "preview": "nuxt preview",
-    "postinstall": "nuxt prepare",
-    "type-check": "golar --noEmit"
+    "dev": "astro dev",
+    "build": "astro build",
+    "preview": "astro preview"
   },
   "dependencies": {
-    "@eslint/js": "^10.0.1",
-    "@nuxt/content": "3.12.0",
-    "@nuxt/eslint": "^1.15.2",
-    "@nuxt/eslint-config": "^1.15.2",
-    "@nuxt/fonts": "^0.14.0",
-    "@nuxt/icon": "^2.2.1",
-    "@nuxtjs/color-mode": "^4.0.0",
+    "@astrojs/cloudflare": "^12.6.13",
+    "@astrojs/vue": "^5.0.12",
     "@tailwindcss/vite": "4.2.2",
-    "@types/luxon": "^3.7.1",
+    "astro": "^5.9.3",
+    "astro-icon": "^1.1.5",
     "es-toolkit": "^1.45.1",
-    "eslint": "^10.1.0",
-    "eslint-plugin-oxlint": "^1.56.0",
-    "eslint-plugin-vue": "^10.8.0",
-    "globals": "^17.4.0",
     "luxon": "^3.7.2",
-    "nuxt": "^4.4.2",
-    "oxfmt": "^0.41.0",
-    "oxlint": "^1.56.0",
-    "oxlint-tsgolint": "^0.17.1",
-    "remark-parse": "^11.0.0",
-    "remark-stringify": "^11.0.0",
-    "strip-markdown": "^6.0.0",
+    "mdast-util-to-string": "^4.0.0",
     "tailwindcss": "4.2.2",
     "tailwindcss-radix-colors": "^2.1.0",
-    "typescript": "^5.9.3",
-    "typescript-eslint": "^8.57.1",
-    "vite": "^8.0.1",
     "vue": "^3.5.30",
-    "vue-router": "^5.0.4",
     "zod": "^4.3.6"
   },
   "devDependencies": {
-    "@golar/vue": "^0.0.16",
-    "debug": "^4.4.3",
-    "golar": "^0.0.16",
-    "knip": "^6.0.1",
-    "parse5": "^8.0.0",
-    "rehype-raw": "^7.0.0",
-    "remark-emoji": "^5.0.2",
-    "remark-gfm": "^4.0.1",
-    "remark-mdc": "^3.10.0",
-    "remark-rehype": "^11.1.2",
-    "unified": "^11.0.5",
-    "unist-util-visit": "^5.1.0",
-    "wrangler": "^4.76.0"
+    "@iconify-json/bi": "^1.2.7",
+    "@iconify-json/lucide": "*",
+    "@iconify-json/mdi": "*",
+    "@iconify-json/proicons": "*",
+    "@types/luxon": "^3.7.1",
+    "typescript": "^5.9.3"
   },
   "engines": {
     "node": ">=24.0.0"
   },
   "packageManager": "pnpm@10.32.1+sha512.a706938f0e89ac1456b6563eab4edf1d1faf3368d1191fc5c59790e96dc918e4456ab2e67d613de1043d2e8c81f87303e6b40d4ffeca9df15ef1ad567348f2be",
   "pnpm": {
-    "ignoredBuiltDependencies": [
-      "@parcel/watcher",
-      "esbuild",
-      "vue-demi"
-    ],
     "onlyBuiltDependencies": [
-      "better-sqlite3"
+      "esbuild",
+      "sharp",
+      "workerd"
     ]
   }
 }
