--- server/utils/cache.ts
@@ -1,24 +1,0 @@
-import { mkdir, readFile, writeFile } from 'node:fs/promises';
-import { join } from 'node:path';
-
-const CACHE_DIR = join(process.cwd(), 'node_modules', '.cache', 'nuxt');
-
-export async function ensureCacheDir(): Promise<void> {
-  await mkdir(CACHE_DIR, { recursive: true });
-}
-
-export async function readCache<T>(filename: string): Promise<T | null> {
-  try {
-    const filePath = join(CACHE_DIR, filename);
-    const data = await readFile(filePath, 'utf-8');
-    return JSON.parse(data) as T;
-  } catch {
-    return null;
-  }
-}
-
-export async function writeCache(filename: string, data: unknown): Promise<void> {
-  await ensureCacheDir();
-  const filePath = join(CACHE_DIR, filename);
-  await writeFile(filePath, JSON.stringify(data, null, 2), 'utf-8');
-}
