+++ src/components/AppFooter.astro
@@ -1,0 +1,13 @@
+---
+import AppSocialLinks from './AppSocialLinks.astro';
+import AppNowPlaying from './AppNowPlaying.vue';
+---
+
+<footer class="mt-16 w-full bg-sage-3 dark:bg-sagedark-2">
+  <div class="mx-auto grid max-w-6xl grid-cols-12 gap-4 px-4 py-12 md:gap-8 lg:gap-12">
+    <div class="col-span-12 flex flex-col gap-12 md:col-span-8 md:col-start-3">
+      <AppNowPlaying client:visible />
+      <AppSocialLinks />
+    </div>
+  </div>
+</footer>
