+++ src/components/DarkModeToggle.vue
@@ -1,0 +1,31 @@
+<script setup lang="ts">
+import { ref, onMounted } from 'vue';
+
+const isDark = ref(false);
+
+onMounted(() => {
+  isDark.value = document.documentElement.classList.contains('dark');
+});
+
+function toggle() {
+  isDark.value = !isDark.value;
+  document.documentElement.classList.toggle('dark', isDark.value);
+  localStorage.setItem('color-mode', isDark.value ? 'dark' : 'light');
+}
+</script>
+
+<template>
+  <button
+    type="button"
+    aria-label="Toggle Dark Mode"
+    class="text-jade-12 dark:text-jadedark-12"
+    @click="toggle"
+  >
+    <svg v-if="isDark" viewBox="0 0 16 16" fill="currentColor" class="size-5">
+      <path d="M6 .278a.77.77 0 0 1 .08.858a7.2 7.2 0 0 0-.878 3.46c0 4.021 3.278 7.277 7.318 7.277q.792-.001 1.533-.16a.79.79 0 0 1 .81.316a.73.73 0 0 1-.031.893A8.35 8.35 0 0 1 8.344 16C3.734 16 0 12.286 0 7.71C0 4.266 2.114 1.312 5.124.06A.75.75 0 0 1 6 .278" />
+    </svg>
+    <svg v-else viewBox="0 0 16 16" fill="currentColor" class="size-5">
+      <path d="M6 .278a.77.77 0 0 1 .08.858a7.2 7.2 0 0 0-.878 3.46c0 4.021 3.278 7.277 7.318 7.277q.792-.001 1.533-.16a.79.79 0 0 1 .81.316a.73.73 0 0 1-.031.893A8.35 8.35 0 0 1 8.344 16C3.734 16 0 12.286 0 7.71C0 4.266 2.114 1.312 5.124.06A.75.75 0 0 1 6 .278M4.858 1.311A7.27 7.27 0 0 0 1.025 7.71c0 4.02 3.279 7.276 7.319 7.276a7.32 7.32 0 0 0 5.205-2.162q-.506.063-1.029.063c-4.61 0-8.343-3.714-8.343-8.29c0-1.167.242-2.278.681-3.286" />
+    </svg>
+  </button>
+</template>
