+++ src/components/HomeRecentPosts.astro
@@ -1,0 +1,28 @@
+---
+import { getCollection } from 'astro:content';
+import PostCard from './PostCard.astro';
+
+const posts = (await getCollection('posts'))
+  .sort((a, b) => b.data.date.getTime() - a.data.date.getTime())
+  .slice(0, 5);
+---
+
+<section class="col-span-12 md:col-span-8 md:col-start-3">
+  <div class="mb-6 flex items-baseline justify-between">
+    <h2 class="text-xs font-medium tracking-widest text-balance text-sage-10 uppercase dark:text-sagedark-10">
+      Recent writing
+    </h2>
+    <a
+      href="/posts"
+      class="text-sm text-sage-10 transition-colors hover:text-jade-11 dark:text-sagedark-10 dark:hover:text-jadedark-11"
+    >
+      View all &rarr;
+    </a>
+  </div>
+
+  <div class="divide-y divide-sage-6 dark:divide-sagedark-6">
+    {posts.map((post) => (
+      <PostCard post={post} />
+    ))}
+  </div>
+</section>
