+++ src/content/posts/managing-ai-config-with-dotfiles.md
