+++ src/pages/posts/[slug].astro
@@ -1,0 +1,57 @@
+---
+import { getCollection, render } from 'astro:content';
+import { DateTime } from 'luxon';
+import BaseLayout from '../../layouts/BaseLayout.astro';
+
+export async function getStaticPaths() {
+  const posts = await getCollection('posts');
+  return posts.map((post) => ({
+    params: { slug: post.id },
+    props: { post },
+  }));
+}
+
+const { post } = Astro.props;
+const { Content, remarkPluginFrontmatter } = await render(post);
+const readingTime = (remarkPluginFrontmatter as { readingTime?: number }).readingTime ?? 1;
+const date = DateTime.fromJSDate(post.data.date);
+---
+
+<BaseLayout title={post.data.title} description={post.data.description}>
+  <div class="prose col-span-12 *:scroll-mt-14 md:col-span-8 md:col-start-3">
+    {post.data.coverImage && (
+      <img
+        src={post.data.coverImage}
+        alt={post.data.title}
+        width="1200"
+        height="800"
+        loading="eager"
+        fetchpriority="high"
+        class="not-prose mb-8 aspect-3/1 w-full rounded-lg object-cover md:-mx-12 md:w-[calc(100%+6rem)] md:max-w-[calc(100%+6rem)]"
+      />
+    )}
+
+    <div class="not-prose mb-8 flex items-center gap-4 text-sm text-sage-10 dark:text-sagedark-10">
+      <time datetime={date.toISODate() ?? undefined}>
+        {date.toLocaleString(DateTime.DATE_MED)}
+      </time>
+      <span>&middot;</span>
+      <span>{readingTime} min read</span>
+    </div>
+
+    <Content class="[&>p>img]:my-8 [&>p>img]:rounded-lg [&>p>img]:md:-mx-12 [&>p>img]:md:my-12 [&>p>img]:md:max-w-[calc(100%+6rem)]" />
+
+    {post.data.tags.length > 0 && (
+      <div class="not-prose mt-12 flex items-center gap-2 text-xs tracking-wide text-sage-10 uppercase dark:text-sagedark-10">
+        {post.data.tags.map((tag) => (
+          <a
+            href={`/posts/tags/${tag}`}
+            class="transition-colors hover:text-jade-11 dark:hover:text-jadedark-11"
+          >
+            #{tag}
+          </a>
+        ))}
+      </div>
+    )}
+  </div>
+</BaseLayout>
